delete not using debug message, refactor viewmodel viewcontroller
