+++ index.html
@@ -25,10 +25,10 @@
       require("./renderer-process/viewcontrollers/view-switch-viewcontroller");
       require("./renderer-process/viewcontrollers/translate-viewcontroller")
       require("./renderer-process/viewcontrollers/search-viewcontroller")
-      require("./renderer-process/search-controller");
       require("./renderer-process/viewcontrollers/favorite-viewcontroller");
       require("./renderer-process/viewcontrollers/read-viewcontroller")
-      require("./renderer-process/subscriber");
+      require("./renderer-process/viewmodels/search-viewmodel");
+      require("./renderer-process/viewmodels/subscribe-viewmodel");
 
       // last things to load
       require("./renderer-process/load-after");
